+++ config/routes.rb
@@ -113,6 +113,10 @@ Rails.application.routes.draw do
           get "completed", to: "projects#completed"
           get "unassigned", to: "projects#unassigned"
         end
+        namespace :regional, path: "regional" do
+          get "in-progress", to: "projects#in_progress"
+          get "completed", to: "projects#completed"
+        end
         namespace :user do
           get "in-progress", to: "projects#in_progress"
           get "completed", to: "projects#completed"