+++ spec/features/projects/users_can_view_a_list_of_regional_projects_spec.rb
@@ -1,0 +1,62 @@
+require "rails_helper"
+
+RSpec.feature "Viewing regional projects" do
+  let(:user) { create(:user, :caseworker) }
+
+  before do
+    sign_in_with_user(user)
+  end
+
+  context "when there are no projects" do
+    scenario "they can view a helpful message" do
+      visit in_progress_regional_projects_path
+
+      expect(page).to have_content(I18n.t("project.table.in_progress.empty"))
+
+      visit completed_regional_projects_path
+
+      expect(page).to have_content(I18n.t("project.table.completed.empty"))
+    end
+  end
+
+  context "when there are projects" do
+    before do
+      mock_successful_api_response_to_create_any_project
+      mock_pre_fetched_api_responses_for_any_establishment_and_trust
+    end
+
+    let!(:completed_regional_project) { create(:conversion_project, urn: 101133, assigned_to_regional_caseworker_team: false, completed_at: Date.yesterday) }
+    let!(:in_progress_regional_project) { create(:conversion_project, urn: 126041, assigned_to_regional_caseworker_team: false, assigned_to: user) }
+
+    let!(:completed_project) { create(:conversion_project, urn: 121583, completed_at: Date.yesterday, assigned_to_regional_caseworker_team: true, assigned_to: user) }
+    let!(:in_progress_project) { create(:conversion_project, urn: 115652, assigned_to_regional_caseworker_team: true, assigned_to: user) }
+
+    scenario "they can view all in progress projects" do
+      visit in_progress_regional_projects_path
+
+      expect(page).to have_content(I18n.t("project.regional.in_progress.title"))
+
+      within("tbody") do
+        expect(page).to_not have_content(in_progress_project.urn)
+        expect(page).to have_content(in_progress_regional_project.urn)
+
+        expect(page).to_not have_content(completed_project.urn)
+        expect(page).to_not have_content(completed_regional_project.urn)
+      end
+    end
+
+    scenario "they can view all completed projects" do
+      visit completed_regional_projects_path
+
+      expect(page).to have_content(I18n.t("project.regional.completed.title"))
+
+      within("tbody") do
+        expect(page).not_to have_content(completed_project.urn)
+        expect(page).to have_content(completed_regional_project.urn)
+
+        expect(page).to_not have_content(in_progress_project.urn)
+        expect(page).to_not have_content(in_progress_regional_project.urn)
+      end
+    end
+  end
+end