+++ CHANGELOG.md
@@ -12,9 +12,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   order has been issued
 - the new project form no longer asks 'Is the school joining a Sponsor trust?'
 
+### Added
+
+- New views at `/projects/regional/in-progress` and
+  `/projects/regional/completed` to show projects NOT assigned to Regional
+  caseworker services
+
 ## [Release 19][release-19]
 
-### Added
+### Added
 
 - a new view at `/projects/all/new` that shows all conversion projects that
   require the new Academy URN to be provided